+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed profile corruption when uninstalling a mod used by multiple profiles
+- Fixed unpacked ROM optional DLC folder check
 
 ## [0.5.0] - 2023-03-02